Do not use output folder when data URI are used, allow to provide a puppeteerConfig file

+++ index.mjs
@@ -13,6 +13,7 @@ let settings = {
   backgroundColor: "white",
   imageAttributes: [],
   mermaidConfig: {},
+  puppeteerConfig: {},
 };
 
 function printSettings() {
@@ -55,10 +56,16 @@ function initialize(options) {
     makeWorkingFilePath("mermaidConfig.json"),
     JSON.stringify(settings.mermaidConfig),
   );
+  fs.writeFileSync(
+    makeWorkingFilePath("puppeteerConfig.json"),
+    JSON.stringify(settings.puppeteerConfig),
+  );
   if (fs.existsSync(settings.outputFolder)) {
     fs.rmSync(settings.outputFolder, { recursive: true, force: true });
   }
-  fs.mkdirSync(settings.outputFolder);
+  if (!settings.useDataUri) {
+    fs.mkdirSync(settings.outputFolder);
+  }
 }
 
 function extractChartTitle(chartDefinition) {
@@ -69,32 +76,45 @@ function extractChartTitle(chartDefinition) {
   }
 }
 
-function mermaidChart(chartDefinition) {
-  const chartId = hashSum(chartDefinition);
+function getChartFileLocation(chartFileName) {
+  if (settings.useDataUri) {
+    return makeWorkingFilePath(chartFileName + ".svg");
+  } else {
+    return makeOutputFilePath(chartFileName + ".svg");
+  }
+}
 
-  try {
+function prepareChartFile(chartDefinition, chartFileName) {
+  if (!fs.existsSync(getChartFileLocation(chartFileName))) {
     fs.writeFileSync(`${settings.workingFolder}/chart.mmd`, chartDefinition);
-
     execSync(
-      `npx -p @mermaid-js/mermaid-cli mmdc -q --backgroundColor ${settings.backgroundColor} --input ${makeWorkingFilePath("chart.mmd")} --output ${makeOutputFilePath(chartId + ".svg")} --configFile ${makeWorkingFilePath("mermaidConfig.json")}`,
+      `npx -p @mermaid-js/mermaid-cli mmdc -q --backgroundColor ${settings.backgroundColor} --input ${makeWorkingFilePath("chart.mmd")} --output ${getChartFileLocation(chartFileName)} --configFile ${makeWorkingFilePath("mermaidConfig.json")} --puppeteerConfigFile ${makeWorkingFilePath("puppeteerConfig.json")}`,
       {
         cwd: "./",
         encoding: "utf-8",
         stdio: "inherit",
       },
     );
+  }
+}
+
+function mermaidChart(chartDefinition) {
+  const chartFileName = hashSum(chartDefinition);
+
+  try {
+    prepareChartFile(chartDefinition, chartFileName);
 
     const chartTitle = extractChartTitle(chartDefinition);
 
     if (settings.useDataUri) {
       const chart = fs.readFileSync(
-        makeOutputFilePath(chartId + ".svg"),
+        getChartFileLocation(chartFileName),
         "binary",
       );
       const chartBase64 = Buffer.from(chart, "binary").toString("base64");
       return `<figure class="mermaid"><img src=\"data:image/svg+xml;base64,${chartBase64}\"${settings.imageAttributes.length ? " " + settings.imageAttributes.join(" ") : ""}/>${chartTitle ? `<figcaption>${chartTitle}</figcaption>` : ""}</figure>`;
     } else {
-      return `<figure class="mermaid"><img src=\"${makeUrlPath(chartId + ".svg")}\"${settings.imageAttributes.length ? " " + settings.imageAttributes.join(" ") : ""}/>${chartTitle ? `<figcaption>${chartTitle}</figcaption>` : ""}</figure>`;
+      return `<figure class="mermaid"><img src=\"${makeUrlPath(chartFileName + ".svg")}\"${settings.imageAttributes.length ? " " + settings.imageAttributes.join(" ") : ""}/>${chartTitle ? `<figcaption>${chartTitle}</figcaption>` : ""}</figure>`;
     }
   } catch (err) {
     console.error(
